refactor participant
