--- system-tests/src/test/java/org/eclipse/edc/samples/transfer/streaming/Participant.java
@@ -1,458 +1,0 @@
-/*
- *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-package org.eclipse.edc.samples.transfer.streaming;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.restassured.specification.RequestSpecification;
-import jakarta.json.Json;
-import jakarta.json.JsonArray;
-import jakarta.json.JsonObject;
-import org.eclipse.edc.jsonld.TitaniumJsonLd;
-import org.eclipse.edc.jsonld.spi.JsonLd;
-import org.eclipse.edc.jsonld.util.JacksonJsonLd;
-import org.eclipse.edc.spi.EdcException;
-import org.eclipse.edc.spi.monitor.ConsoleMonitor;
-
-import java.net.URI;
-import java.time.Duration;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.UUID;
-import java.util.concurrent.atomic.AtomicReference;
-
-import static io.restassured.RestAssured.given;
-import static io.restassured.http.ContentType.JSON;
-import static jakarta.json.Json.createArrayBuilder;
-import static jakarta.json.Json.createObjectBuilder;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.awaitility.Awaitility.await;
-import static org.eclipse.edc.connector.controlplane.contract.spi.types.negotiation.ContractNegotiationStates.FINALIZED;
-import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.CONTEXT;
-import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.ID;
-import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.TYPE;
-import static org.eclipse.edc.jsonld.spi.PropertyAndTypeNames.DCAT_DATASET_ATTRIBUTE;
-import static org.eclipse.edc.jsonld.spi.PropertyAndTypeNames.ODRL_ASSIGNER_ATTRIBUTE;
-import static org.eclipse.edc.jsonld.spi.PropertyAndTypeNames.ODRL_POLICY_ATTRIBUTE;
-import static org.eclipse.edc.jsonld.spi.PropertyAndTypeNames.ODRL_TARGET_ATTRIBUTE;
-import static org.eclipse.edc.spi.constants.CoreConstants.EDC_NAMESPACE;
-import static org.eclipse.edc.spi.constants.CoreConstants.EDC_PREFIX;
-
-/**
- * Essentially a wrapper around the management API enabling to test interactions with other participants, eg. catalog, transfer...
- */
-public class Participant {
-
-    private static final String DSP_PROTOCOL = "dataspace-protocol-http";
-    private static final Duration TIMEOUT = Duration.ofSeconds(30);
-
-    protected String id;
-    protected String name;
-    protected Endpoint managementEndpoint;
-    protected Endpoint protocolEndpoint;
-    protected Endpoint controlEndpoint;
-    protected JsonLd jsonLd;
-    protected ObjectMapper objectMapper;
-
-    protected Participant() {
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void registerDataPlane(List<String> sourceTypes, List<String> destinationTypes, List<Object> transferTypes) {
-        var jsonObject = Json.createObjectBuilder()
-                .add(CONTEXT, createObjectBuilder().add(EDC_PREFIX, EDC_NAMESPACE))
-                .add(ID, UUID.randomUUID().toString())
-                .add(EDC_NAMESPACE + "url", controlEndpoint.url + "/transfer")
-                .add(EDC_NAMESPACE + "allowedSourceTypes", createArrayBuilder(sourceTypes))
-                .add(EDC_NAMESPACE + "allowedDestTypes", createArrayBuilder(destinationTypes))
-                .add(EDC_NAMESPACE + "allowedTransferTypes", createArrayBuilder(transferTypes))
-                .build();
-
-        managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .body(jsonObject.toString())
-                .when()
-                .post("/v2/dataplanes")
-                .then()
-                .statusCode(200);
-    }
-
-    public String createAsset(String requestBody) {
-        return managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .body(requestBody)
-                .when()
-                .post("/v3/assets")
-                .then()
-                .statusCode(200)
-                .contentType(JSON)
-                .extract().jsonPath().getString(ID);
-    }
-
-    public String createPolicyDefinition(String requestBody) {
-        return managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .body(requestBody)
-                .when()
-                .post("/v2/policydefinitions")
-                .then()
-                .statusCode(200)
-                .contentType(JSON)
-                .extract().jsonPath().getString(ID);
-    }
-
-    public String createContractDefinition(String requestBody) {
-        return managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .body(requestBody)
-                .when()
-                .post("/v2/contractdefinitions")
-                .then()
-                .statusCode(200)
-                .extract().jsonPath().getString(ID);
-    }
-
-    /**
-     * Request provider catalog.
-     *
-     * @param provider data provider
-     * @return list of {@link org.eclipse.edc.catalog.spi.Dataset}.
-     */
-    public JsonArray getCatalogDatasets(Participant provider) {
-        var datasetReference = new AtomicReference<JsonArray>();
-        var requestBody = createObjectBuilder()
-                .add(CONTEXT, createObjectBuilder().add(EDC_PREFIX, EDC_NAMESPACE))
-                .add(TYPE, "CatalogRequest")
-                .add("counterPartyAddress", provider.protocolEndpoint.url.toString())
-                .add("protocol", DSP_PROTOCOL)
-                .build();
-
-        await().atMost(TIMEOUT).untilAsserted(() -> {
-            var response = managementEndpoint.baseRequest()
-                    .contentType(JSON)
-                    .when()
-                    .body(requestBody)
-                    .post("/v2/catalog/request")
-                    .then()
-                    .log().ifValidationFails()
-                    .statusCode(200)
-                    .extract().body().asString();
-
-            var responseBody = objectMapper.readValue(response, JsonObject.class);
-
-            var catalog = jsonLd.expand(responseBody).orElseThrow(f -> new EdcException(f.getFailureDetail()));
-
-            var datasets = catalog.getJsonArray(DCAT_DATASET_ATTRIBUTE);
-            assertThat(datasets).hasSizeGreaterThan(0);
-
-            datasetReference.set(datasets);
-        });
-
-        return datasetReference.get();
-    }
-
-    /**
-     * Get first {@link org.eclipse.edc.catalog.spi.Dataset} from provider matching the given asset id.
-     *
-     * @param provider data provider
-     * @param assetId  asset id
-     * @return dataset.
-     */
-    public JsonObject getDatasetForAsset(Participant provider, String assetId) {
-        var datasetReference = new AtomicReference<JsonObject>();
-        var requestBody = createObjectBuilder()
-                .add(CONTEXT, createObjectBuilder().add(EDC_PREFIX, EDC_NAMESPACE))
-                .add(TYPE, "DatasetRequest")
-                .add(ID, assetId)
-                .add("counterPartyAddress", provider.protocolEndpoint.url.toString())
-                .add("protocol", DSP_PROTOCOL)
-                .build();
-
-        await().atMost(TIMEOUT).untilAsserted(() -> {
-            var response = managementEndpoint.baseRequest()
-                    .contentType(JSON)
-                    .when()
-                    .body(requestBody)
-                    .post("/v2/catalog/dataset/request")
-                    .then()
-                    .log().ifValidationFails()
-                    .statusCode(200)
-                    .extract().body().asString();
-
-            var compacted = objectMapper.readValue(response, JsonObject.class);
-
-            var dataset = jsonLd.expand(compacted).orElseThrow(f -> new EdcException(f.getFailureDetail()));
-
-            datasetReference.set(dataset);
-        });
-
-        return datasetReference.get();
-    }
-
-    /**
-     * Initiate negotiation with a provider.
-     *
-     * @param provider data provider
-     * @param offer    the contract offer
-     * @return id of the contract agreement.
-     */
-    public String negotiateContract(Participant provider, JsonObject offer) {
-        var requestBody = createObjectBuilder()
-                .add(CONTEXT, createObjectBuilder().add(EDC_PREFIX, EDC_NAMESPACE))
-                .add(TYPE, "ContractRequestDto")
-                .add("providerId", provider.id)
-                .add("counterPartyAddress", provider.protocolEndpoint.url.toString())
-                .add("protocol", DSP_PROTOCOL)
-                .add("policy", jsonLd.compact(offer).getContent())
-                .build();
-
-        var negotiationId = managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .body(requestBody)
-                .when()
-                .post("/v2/contractnegotiations")
-                .then()
-                .log().ifValidationFails()
-                .statusCode(200)
-                .extract().body().jsonPath().getString(ID);
-
-        await().atMost(TIMEOUT).untilAsserted(() -> {
-            var state = getContractNegotiationState(negotiationId);
-            assertThat(state).isEqualTo(FINALIZED.name());
-        });
-
-        return getContractAgreementId(negotiationId);
-    }
-
-    /**
-     * Initiate data transfer.
-     *
-     * @param provider            data provider
-     * @param contractAgreementId contract agreement id
-     * @param assetId             asset id
-     * @param privateProperties   private properties
-     * @param destination         data destination address
-     * @param transferType        transfer type
-     * @return id of the transfer process.
-     */
-    public String initiateTransfer(Participant provider, String contractAgreementId, String assetId, JsonObject privateProperties, JsonObject destination, String transferType) {
-        var builder = createObjectBuilder()
-                .add(CONTEXT, createObjectBuilder().add(EDC_PREFIX, EDC_NAMESPACE))
-                .add(TYPE, "TransferRequest")
-                .add("protocol", DSP_PROTOCOL)
-                .add("assetId", assetId)
-                .add("contractId", contractAgreementId)
-                .add("connectorId", provider.id)
-                .add("counterPartyAddress", provider.protocolEndpoint.url.toString())
-                .add("privateProperties", privateProperties)
-                .add("transferType", transferType);
-
-        if (destination != null) {
-            builder.add("dataDestination", destination);
-        }
-
-        var requestBody = builder.build();
-
-        return managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .body(requestBody)
-                .when()
-                .post("/v2/transferprocesses")
-                .then()
-                .log().ifError()
-                .statusCode(200)
-                .extract().body().jsonPath().getString(ID);
-    }
-
-    /**
-     * Request a provider asset:
-     * - retrieves the contract definition associated with the asset,
-     * - handles the contract negotiation,
-     * - initiate the data transfer.
-     *
-     * @param provider          data provider
-     * @param assetId           asset id
-     * @param privateProperties private properties of the data request
-     * @param destination       data destination
-     * @param transferType      transfer type
-     * @return transfer process id.
-     */
-    public String requestAsset(Participant provider, String assetId, JsonObject privateProperties, JsonObject destination, String transferType) {
-        var offer = getOfferForAsset(provider, assetId);
-
-        var contractAgreementId = negotiateContract(provider, offer);
-        var transferProcessId = initiateTransfer(provider, contractAgreementId, assetId, privateProperties, destination, transferType);
-        assertThat(transferProcessId).isNotNull();
-        return transferProcessId;
-    }
-
-    private JsonObject getOfferForAsset(Participant provider, String assetId) {
-        var dataset = getDatasetForAsset(provider, assetId);
-        var policy = dataset.getJsonArray(ODRL_POLICY_ATTRIBUTE).get(0).asJsonObject();
-        return createObjectBuilder(policy)
-                .add(ODRL_ASSIGNER_ATTRIBUTE, createObjectBuilder().add(ID, provider.id))
-                .add(ODRL_TARGET_ATTRIBUTE, createObjectBuilder().add(ID, dataset.get(ID)))
-                .build();
-    }
-
-    /**
-     * Get current state of a transfer process.
-     *
-     * @param id transfer process id
-     * @return state of the transfer process.
-     */
-    public String getTransferProcessState(String id) {
-        return managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .when()
-                .get("/v2/transferprocesses/{id}/state", id)
-                .then()
-                .statusCode(200)
-                .extract().body().jsonPath().getString("state");
-    }
-
-    private String getContractNegotiationState(String id) {
-        return managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .when()
-                .get("/v2/contractnegotiations/{id}/state", id)
-                .then()
-                .statusCode(200)
-                .extract().body().jsonPath().getString("state");
-    }
-
-
-    private String getContractAgreementId(String negotiationId) {
-        var contractAgreementIdAtomic = new AtomicReference<String>();
-
-        await().atMost(TIMEOUT).untilAsserted(() -> {
-            var agreementId = getContractNegotiationField(negotiationId, "contractAgreementId");
-            assertThat(agreementId).isNotNull().isInstanceOf(String.class);
-
-            contractAgreementIdAtomic.set(agreementId);
-        });
-
-        var contractAgreementId = contractAgreementIdAtomic.get();
-        assertThat(id).isNotEmpty();
-        return contractAgreementId;
-    }
-
-    private String getContractNegotiationField(String negotiationId, String fieldName) {
-        return managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .when()
-                .get("/v2/contractnegotiations/{id}", negotiationId)
-                .then()
-                .statusCode(200)
-                .extract().body().jsonPath()
-                .getString(fieldName);
-    }
-
-    /**
-     * Represent an endpoint exposed by a {@link Participant}.
-     */
-    public static class Endpoint {
-        private final URI url;
-        private final Map<String, String> headers;
-
-        public Endpoint(URI url) {
-            this.url = url;
-            this.headers = new HashMap<>();
-        }
-
-        public Endpoint(URI url, Map<String, String> headers) {
-            this.url = url;
-            this.headers = headers;
-        }
-
-        public RequestSpecification baseRequest() {
-            return given().baseUri(url.toString()).headers(headers);
-        }
-
-        public URI getUrl() {
-            return url;
-        }
-    }
-
-    public static class Builder<P extends Participant, B extends Builder<P, B>> {
-        protected final P participant;
-
-        protected Builder(P participant) {
-            this.participant = participant;
-        }
-
-        public static <B extends Builder<Participant, B>> Builder<Participant, B> newInstance() {
-            return new Builder<>(new Participant());
-        }
-
-        public B id(String id) {
-            participant.id = id;
-            return self();
-        }
-
-        public B name(String name) {
-            participant.name = name;
-            return self();
-        }
-
-        public B managementEndpoint(Endpoint managementEndpoint) {
-            participant.managementEndpoint = managementEndpoint;
-            return self();
-        }
-
-        public B protocolEndpoint(Endpoint protocolEndpoint) {
-            participant.protocolEndpoint = protocolEndpoint;
-            return self();
-        }
-
-        public B controlEndpoint(Endpoint controlEndpoint) {
-            participant.controlEndpoint = controlEndpoint;
-            return self();
-        }
-
-        public B jsonLd(JsonLd jsonLd) {
-            participant.jsonLd = jsonLd;
-            return self();
-        }
-
-        public B objectMapper(ObjectMapper objectMapper) {
-            participant.objectMapper = objectMapper;
-            return self();
-        }
-
-        public Participant build() {
-            Objects.requireNonNull(participant.id, "id");
-            Objects.requireNonNull(participant.name, "name");
-            Objects.requireNonNull(participant.managementEndpoint, "managementEndpoint");
-            Objects.requireNonNull(participant.protocolEndpoint, "protocolEndpoint");
-            if (participant.jsonLd == null) {
-                participant.jsonLd = new TitaniumJsonLd(new ConsoleMonitor());
-            }
-            if (participant.objectMapper == null) {
-                participant.objectMapper = JacksonJsonLd.createObjectMapper();
-            }
-            return participant;
-        }
-
-        @SuppressWarnings("unchecked")
-        private B self() {
-            return (B) this;
-        }
-    }
-}
