+++ system-tests/src/test/java/org/eclipse/edc/samples/transfer/streaming/Streaming02KafkaToHttpTest.java
@@ -56,19 +56,19 @@ public class Streaming02KafkaToHttpTest {
     private static final String MAX_DURATION = "PT30S";
     private static final String SAMPLE_FOLDER = "transfer/streaming/streaming-02-kafka-to-http";
     private static final Duration TIMEOUT = Duration.ofSeconds(30);
-    private static final Participant PROVIDER = Participant.Builder.newInstance()
+    private static final StreamingParticipant PROVIDER = StreamingParticipant.Builder.newStreamingInstance()
             .name("provider")
             .id("provider")
-            .managementEndpoint(new Participant.Endpoint(URI.create("http://localhost:18181/management")))
-            .protocolEndpoint(new Participant.Endpoint(URI.create("http://localhost:18182/protocol")))
-            .controlEndpoint(new Participant.Endpoint(URI.create("http://localhost:18183/control")))
+            .managementEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:18181/management")))
+            .protocolEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:18182/protocol")))
+            .controlEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:18183/control")))
             .build();
-    private static final Participant CONSUMER = Participant.Builder.newInstance()
+    private static final StreamingParticipant CONSUMER = StreamingParticipant.Builder.newStreamingInstance()
             .name("consumer")
             .id("consumer")
-            .managementEndpoint(new Participant.Endpoint(URI.create("http://localhost:28181/management")))
-            .protocolEndpoint(new Participant.Endpoint(URI.create("http://localhost:28182/protocol")))
-            .controlEndpoint(new Participant.Endpoint(URI.create("http://localhost:28183/control")))
+            .managementEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:28181/management")))
+            .protocolEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:28182/protocol")))
+            .controlEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:28183/control")))
             .build();
 
     @Container
@@ -124,8 +124,10 @@ public class Streaming02KafkaToHttpTest {
                 .add("baseUrl", "http://localhost:" + httpReceiverPort)
                 .build();
 
-        var transferProcessId = CONSUMER.requestAsset(PROVIDER, "kafka-stream-asset",
-                Json.createObjectBuilder().build(), destination, "HttpData-PUSH");
+        var transferProcessId = CONSUMER.requestAssetFrom("kafka-stream-asset", PROVIDER)
+                .withDestination(destination)
+                .withTransferType("HttpData-PUSH")
+                .execute();
 
         await().atMost(TIMEOUT).untilAsserted(() -> {
             var state = CONSUMER.getTransferProcessState(transferProcessId);
