+++ system-tests/src/test/java/org/eclipse/edc/samples/transfer/streaming/Streaming03KafkaToKafkaTest.java
@@ -84,19 +84,19 @@ public class Streaming03KafkaToKafkaTest {
     private static final Path SAMPLE_FOLDER = Path.of("transfer", "streaming", SAMPLE_NAME);
     private static final Path RUNTIME_PATH = SAMPLE_FOLDER.resolve(RUNTIME_NAME);
     private static final Duration TIMEOUT = Duration.ofSeconds(60);
-    private static final Participant PROVIDER = Participant.Builder.newInstance()
+    private static final StreamingParticipant PROVIDER = StreamingParticipant.Builder.newStreamingInstance()
             .name("provider")
             .id("provider")
-            .managementEndpoint(new Participant.Endpoint(URI.create("http://localhost:18181/management")))
-            .protocolEndpoint(new Participant.Endpoint(URI.create("http://localhost:18182/protocol")))
-            .controlEndpoint(new Participant.Endpoint(URI.create("http://localhost:18183/control")))
+            .managementEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:18181/management")))
+            .protocolEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:18182/protocol")))
+            .controlEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:18183/control")))
             .build();
-    private static final Participant CONSUMER = Participant.Builder.newInstance()
+    private static final StreamingParticipant CONSUMER = StreamingParticipant.Builder.newStreamingInstance()
             .name("consumer")
             .id("consumer")
-            .managementEndpoint(new Participant.Endpoint(URI.create("http://localhost:28181/management")))
-            .protocolEndpoint(new Participant.Endpoint(URI.create("http://localhost:28182/protocol")))
-            .controlEndpoint(new Participant.Endpoint(URI.create("http://localhost:28183/control")))
+            .managementEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:28181/management")))
+            .protocolEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:28182/protocol")))
+            .controlEndpoint(new StreamingParticipant.Endpoint(URI.create("http://localhost:28183/control")))
             .build();
     private static final String GROUP_ID = "group_id";
 
@@ -155,7 +155,11 @@ public class Streaming03KafkaToKafkaTest {
         var transferProcessPrivateProperties = Json.createObjectBuilder()
                 .add("receiverHttpEndpoint", "http://localhost:" + httpReceiverPort)
                 .build();
-        var transferProcessId = CONSUMER.requestAsset(PROVIDER, "kafka-stream-asset", transferProcessPrivateProperties, destination, "KafkaBroker-PULL");
+        var transferProcessId = CONSUMER.requestAssetFrom("kafka-stream-asset", PROVIDER)
+                .withPrivateProperties(transferProcessPrivateProperties)
+                .withDestination(destination)
+                .withTransferType("KafkaBroker-PULL")
+                .execute();
 
         await().atMost(TIMEOUT).untilAsserted(() -> {
             var state = CONSUMER.getTransferProcessState(transferProcessId);
